[sync] [Core] Move MRV1 `late_interaction_runner.py` out of MRV2 subtree (#48014)
Moved LateInteractionRunner to aphrodite/v1/pool/late_interaction_runner.py and updated the test and gpu_model_runner imports to use aphrodite.v1.pool.late_interaction_runner. Verified no stale imports remain.

Upstream-vLLM: a5d19cbb95872c4b426c06735733568542fa33db
